Refactor CRI packages
