+++ pkg/cri/config/config_unix.go
@@ -21,7 +21,7 @@ package config
 import (
 	"github.com/BurntSushi/toml"
 	"github.com/containerd/containerd"
-	"github.com/containerd/containerd/pkg/streaming"
+	"github.com/containerd/containerd/pkg/cri/streaming"
 )
 
 // DefaultConfig returns default configurations of cri plugin.
