+++ pkg/cri/server/sandbox_run_linux_test.go
@@ -28,8 +28,8 @@ import (
 	"github.com/stretchr/testify/require"
 	runtime "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 
-	"github.com/containerd/containerd/pkg/annotations"
 	"github.com/containerd/containerd/pkg/containerd/opts"
+	"github.com/containerd/containerd/pkg/cri/annotations"
 	ostesting "github.com/containerd/containerd/pkg/os/testing"
 )
