+++ pkg/cri/server/version.go
@@ -21,7 +21,7 @@ import (
 	"golang.org/x/net/context"
 	runtime "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 
-	"github.com/containerd/containerd/pkg/constants"
+	"github.com/containerd/containerd/pkg/cri/constants"
 )
 
 const (
